+++ Assets/DeviceID.txt
@@ -1,1 +1,1 @@
-﻿2
+﻿1
