+++ Assets/Plugins/SteamVR/Scripts/SteamVR_ControllerManager.cs
@@ -9,8 +9,7 @@ using Valve.VR;
 
 public class SteamVR_ControllerManager : MonoBehaviour
 {
-    public static SteamVR_ControllerManager instance;
-    public GameObject left, right;
+	public GameObject left, right;
 
 	[Tooltip("Populate with objects you want to assign to additional controllers")]
 	public GameObject[] objects;
@@ -57,17 +56,10 @@ public class SteamVR_ControllerManager : MonoBehaviour
 
 	void Awake()
 	{
-        if (instance != null)
-        {
-            DestroyImmediate(gameObject);
-            return;
-        }
-        instance = this;
-        DontDestroyOnLoad(gameObject);
-        UpdateTargets();
+		UpdateTargets();
 	}
 
-    SteamVR_ControllerManager()
+	SteamVR_ControllerManager()
 	{
 		inputFocusAction = SteamVR_Events.InputFocusAction(OnInputFocus);
 		deviceConnectedAction = SteamVR_Events.DeviceConnectedAction(OnDeviceConnected);
