--- Assets/Scenes/Scene2.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: be70b0222c97c53449023689cad7ff6a
-timeCreated: 1506493399
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
