--- Assets/Scenes/Scene3.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5a2b820cb6127684c81deb1ebfcf4b20
-timeCreated: 1506493399
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
