+++ Assets/Scripts/DummysDataControl.cs
@@ -4,15 +4,4 @@ using UnityEngine;
 
 public class DummysDataControl : MonoBehaviour
 {
-    public static DummysDataControl instance;
-    void Awake()
-    {
-        if (instance != null)
-        {
-            DestroyImmediate(gameObject);
-            return;
-        }
-        instance = this;
-        //DontDestroyOnLoad(gameObject);
-    }
 }
