+++ Assets/Scripts/MainControl.cs
@@ -13,7 +13,7 @@ public class MainControl : MonoBehaviour
 
     public GameObject gameMasterControl;
     public GameObject cameraRig;
-    public int nowSceneID = -1;
+
     public PlayerViewsControl playerViewsControl;
 
     void Awake()
@@ -24,13 +24,10 @@ public class MainControl : MonoBehaviour
             return;
         }
         instance = this;
-        //DontDestroyOnLoad(gameObject);
     }
 
     void Start ()
     {
-        nowSceneID = 1;
-
         if (PhotonNetwork.inRoom)
         {
             if(PhotonControl.instance.deviceID == 0)
